- Use dictionary instead of list for feature loopup
- Use a set instead of a list for ids

Before:

```
⇒  time python  sandbox.py
python sandbox.py  39,61s user 0,21s system 99% cpu 39,822 total
```

After:

```
⇒  time python  sandbox.py
python sandbox.py  4,85s user 0,09s system 99% cpu 4,947 total
```