@reckart 2018-09-16
I don't understand the python versioning scheme... for me, a version increase is done when a release is being performed. Is merging a PR automatically a release? What happens when we have more than 26 PRs - does it go "1.1.1aa" then?